basic function generics
Summary:
In order to track information flow through functions with generics, I updated the `subtype` function to more smartly deal wit `Tgeneric`. In particular, we should allow `Tgeneric` to be a sub or super type of anything else and generate flows accordingly.

The subtype case is pretty easy. If a generic is a subtype of something else, then the generic's policy guards the other type (using `add_dependencies`.

The supertype is a bit more tricky. I implementing it by getting the `policy_occurrences` for the subtype of the generic and then creating flows to and from them as needed.

This system should be sound, although it is a bit imprecise. Because the generic only has a single policy, it can cause things to get a bit messed up. For example, if a function type is passed in to a generic position, the resulting function is a lot more constrained (see example in test case). I don't think this is a common usage pattern, so I don't anticipate it to be a huge problem.

Reviewed By: mpu

Differential Revision: D24018107

fbshipit-source-id: 3a493382a0a127ad1a1161a58753270f82d74720