Support constrained-lval for SetOp
Summary:
By tweaking the contracts around the Elem operations, we can make it both easier to understand and also support SetOp (which previously still needed an Lval). The new rules are:
1. The callee of Elem can't return an lval to a persistent string or array-like
2. The caller of Elem can't change the type it points to.

Rule 1 allows us to use Elem for Dim, since it never escalates.

We can rewrite IncDec and SetOp to use a temp TypedValue to honor rule 2, allowing us to use Elem for strictly more cases than we did before. If we update the temp's type, we do a set.

Differential Revision: D24072945

fbshipit-source-id: 7fdc11e7b1e16e961b7b4bb73da8ade4a09c2d90