Reset local typing environment for ExprTrees
Summary:
Expression trees should not inherit the local typing environment. So the following should throw an error:

```
function test(): void {
  $x = 1;

  Code`$x + 1`;
}
```

As the expression tree should not inherit the local type for `$x`. This diff resets the locals in the typing environment for expression trees.

This diff does not account for splicing yet.

Reviewed By: Wilfred

Differential Revision: D23718691

fbshipit-source-id: 7589ea088278cabd56ea634520c39aa205f3469d