Flow typing for Spliced Expressions
Summary:
Spliced Expressions within Expression Trees are really interesting to type because they are typed in a different environment than the surrounding Expression Tree.

```
$x = new Foo();
Code`() ==> {
    __splice__($x->x);
    __splice__($x->reset());
  }
`
```
Should be typed as if the code were:
```
$x = new Foo();
$1 = $x->x;
$2 = $x->reset();
Code`() ==> {
    __splice__($1);
    __splice__($2);
  }
`
```

Where we assume that the execution of the spliced expressions happen before the expression tree is created.

In order to type this, we extract all of the spliced expressions from the expression tree. We then type these expressions with the external environment. We then reset the environment to type the expression tree itself.

When we type the extracted spliced expressions, I store the typed expressions and the inferred types in a map. This is so that when we are typing the `ET_Splice` node, we don't need to do anything with the environment, we can just grab the already inferred and computer types and use them directly.

In order to match the computer types and the `ET_Splice` nodes, when I extract the spliced expressions, I also go and replace them with a placeholder `Id` node that will be used as the key to the map.

Note: Actually typing the splices are not quite right. Right now, I just pass back the type of the spliced expression, but it should expect that the expression is a `ExprTree<X, inferredType>` and we should unwrap the `inferredType` and give that back.

Reviewed By: shayne-fletcher

Differential Revision: D23922161

fbshipit-source-id: 1871e1dda50ff7fc7d4249124ac31c9531d397bf