Replace remaining uses of get_name("", ...)
Summary:
The `get_name` method in the direct decl parser has a lot of responsibilities. It converts simple identifiers (like method names) from `Node` to `Id`, it converts `QualifiedName`s to strings containing backslashes, and elaborates names in a given namespace (which is always either the current namespace or the empty string).

I think we ought to split it into a few helpers that are clearer about their responsibilities (passing the empty string as the current namespace strikes me as a code smell). This diff removes the remaining call sites where we pass an empty string.

Reviewed By: shiqicao

Differential Revision: D24084534

fbshipit-source-id: 9098b2542711ce28b2c153f515e5dcf75eae75d0