Pull qualified_name_from_parts out into method
Summary:
This was locally defined inside `get_name`, but I think both functions are a bit easier to read if we pull this helper out and make it a method too.

This change also counts the length of the name instead of estimating it, to avoid arena garbage from unused String capacity or reallocated Strings.

Reviewed By: shiqicao

Differential Revision: D24084535

fbshipit-source-id: 6425e495664b71282d1dd4ea5f6c46c116b43835