Replace get_name with elaborate_name
Summary:
Now that get_name is nearly always invoked with the current namespace, replace it with a helper `elaborate_name`, which uses the current namespace implicitly.

The invocation of `get_name` inside `function_into_ty` would pass an empty namespace if one was passed to `function_into_ty`, so this diff changes that helper to take a bool `is_method` parameter instead, which controls whether or not we elaborate the name. The helper is also renamed to `function_into_ty`, since "into" is idiomatically reserved for functions which take their (non-Copy) parameter by value. That used to be the case for this helper, but `Node` is now `Copy`.

Reviewed By: shiqicao

Differential Revision: D24084533

fbshipit-source-id: c10900983abae8d5baa946e0fba48094b4313fe9