Replace some uses of get_name("", ...)
Summary:
The `get_name` method in the direct decl parser has a lot of responsibilities. It converts simple identifiers (like method names) from `Node` to `Id`, it converts `QualifiedName`s to strings containing backslashes, and elaborates names in a given namespace (which is always either the current namespace or the empty string).

I think we ought to split it into a few helpers that are clearer about their responsibilities (passing the empty string as the current namespace strikes me as a code smell).

Reviewed By: shiqicao

Differential Revision: D24084536

fbshipit-source-id: 8000152cb01f35487372942e736667a781b36c4d