Fix appends in preg implementation
Summary:
Turning on dict/darray append notices was a huge regression (around 3%), even when we limit these notices only to cases where m_nextKI != m_size.

It turns out that all of the notices come from users of the preg APIs, which store sub-pattern results in arrays indexed by both their string names and integer indices. We can rewrite all of this code to use simple known-index sets instead. (For the arrays which I've added the isVectorData assertion, it would be even better to use a vec/varray, but oh well...)

Reviewed By: ricklavoie

Differential Revision: D24099880

fbshipit-source-id: 7ef09ce3548e302486cb36eada2c8ec03f99eeec