Port hh_fanout CLI to Cmdliner
Summary:
Change the CLI to use `Cmdliner` rather than `Command` from the `Core` library, this is in an effort to eliminate `Core` as a dependency.

The change is mainly coesmetic. The help pages look different but contain the same information as before. Cmdliner by default displays help pages with a pager which I think is better than dumping to STDOUT, so I changed that too. If need be there is an option to change this.

I removed the `help` subcommand because `Cmdliner` enables `--help` to be used with `hh_fanout` and also with `hh_fanout COMMAND --help` automatically. The `version` command is also removed because it was previously displaying nonsensical information rather than an actual version.

Reviewed By: vsiles

Differential Revision: D24081906

fbshipit-source-id: cbdc157e5df7613120a3790c8762ddc2081b955b