Avoid copy for fully qualified names
Summary: We currently always allocate qualified names in our arena, but there's no reason to do so if they're fully qualified in the original source text, since we don't need to prepend a namespace to them. We can directly reference the string in the source text instead of copying it.

Differential Revision: D24091590

fbshipit-source-id: bae81443480fef3a97d1dd8136f0c1369007b981