Fix RepoOption caching
Summary:
We have a number of optimizations to make looking up the appropriate
RepoOption for a file quicker. Unfortunately these optimizations don't
quite work correctly. We'll cache the data, but the cache lookup is
incorrect. Therefore we'll fallback to the slower paths every
time. The problem is that when we validate the cached entry, we check
if the path for the .hhvmconfig.hdf file is a prefix of the current
file. However, we don't strip off the filename. For example, we check
if /www/.hhvmconfig.hdf is a prefix is /www/a/b/c/d.php, which it is
not. Instead we should be stripping off the filename and then doing
the prefix check (/www/ is a prefix of /www/a/b/c/d.php).

Reviewed By: paulbiss

Differential Revision: D24107174

fbshipit-source-id: d886e0fbd7dcbf2ff055d18515935242f14526a0