Unit prefetcher improvements
Summary:
Improve the unit prefetcher in two ways:

First, rework the heuristic for suppressing duplicate prefetch
requests. The current mechanism attempts to keep a "version" counter
per path. We only attempt a prefetch if the version for the path is
less than the "global version". The global version is incremented
whenever a file change is detected in the normal Unit load path. This
works, but it has some deficiences. The most problematic aspect is
that it works poorly in situations where source changes are done using
symlinks. In that case, we may never detect a changed unit, only a
brand new one. The global version is never bumped, so we stop
prefetching all together.

Replace this with a somewhat simpler heuristic. Instead of recording a
version per path, record the timestamp of the last prefetch
attempt. When attempting to prefetch via a particular path, check the
stored timestamp (if any). Only attempt the prefetch if the last
attempt was more than N seconds in the past (right now 15). This
heuristic is a bit more coarse (it won't detect units which change
quickly after we prefetch them), but has less failure modes (it will
ensure that symlink changes won't freeze it).

The second improvement is to move most of the work on the enqueue path
into the worker thread. This keeps the impact on the request threads
to a bare minimum and pushes it onto the worker threads which can
absorb the extra work more easily. This is not straight-forward
because much of the path normalization logic can use per-request
state. However, for a very common set of conditions (absolute path,
certain runtime options), we can safely defer the path
canonicalization onto the worker threads. Luckily these conditions are
the typical case. Path normalization is relatively expensive, as it
may involve stat-ing and realpath-ing the path. Resolving symlinks can
always be done on the worker thread.

Reviewed By: paulbiss

Differential Revision: D24107581

fbshipit-source-id: f4926dc3d57493a8c71801f8eb2612c8fbe3253b