Print capabilities as sets when possible
Summary:
Under normal circumstances each coeffect should be desugarable
into a set of capability types, since the underlying type is of the form:

  Capability1 & ... & CapabilityN

with the special case of `mixed` (or equivalently an *empty* intersection).

In the unusual scenarios, when the internals of the coeffects system
are abused (but unlikely in a realistic use case), one could end up with
unions (due to usage of optional or like- types); therefore, fall back
to the default type printing in such cases.

Example error message as of this didd:
```
Typing[4390] This call is not allowed because its coeffects are incompatible with the context [1]
-> From this declaration, the context of this function body provides the capability set {Rx} [2]
-> But the function being called requires the capability set {NonDet, RxLocal, Throwing<mixed>} [3]

hphp/hack/test/typecheck/coeffects/rx_no_unsafe.php:12:3
     4 | // see hhi/coeffect/contexts.hhi for mapping to capabilities
     5 |
[3]  6 | function non_rx_fn()[non_rx]: void {
     7 |   rx_fn();
     8 | }
     9 |
[2] 10 | function rx_fn()[rx]: void {
    11 |   pure_fn();
[1] 12 |   non_rx_fn(); // error (non_rx context requires additional capabilities)
    13 | }
```

Reviewed By: vassilmladenov

Differential Revision: D24074700

fbshipit-source-id: 079b9479dd2fee938b990125b25df003912a2ec3