zstd for hackc compiler
Summary:
this does no tuning of the zstd compression; ideally we land this and tune after validating production behavior.

no dictionary is used because there is exactly one compressed value.
the shift in distribution is significant.
p95 of cpu and time spent firing up the compiler using zstd is less than the p5 of gzip.

Reviewed By: paulbiss

Differential Revision: D20649569

fbshipit-source-id: 414c8018b5a3405f9af15d0975fe0845b96b9250