Error when reified method + __DynamicallyCallable attribute
Summary:
You are unable to call a reified function or method via `dynamic_fun`, `dynamic_class_meth`, `dynamic_fun_force` and `dynamic_class_meth_force`. You get an error when you attempt to do so. However, since this error is a runtime error, users may still want to annotate reified functions and methods with `__DynamicallyCallable`.

Just throw a Hack error if they attempt to mark a reified function or method with that attribute.

Reviewed By: Wilfred

Differential Revision: D24065373

fbshipit-source-id: 8cc0a56cd0f27b2ba540c3a26c229b9d85a2a7c5