Special error for abstract function pointer
Summary:
Currently when we throw an error on a first class function pointer to an abstract method, we give the error: `Cannot call "Foo::bar()"; it is abstract`, which is the error we throw when we have a call.

Instead, let's throw a more specific error for function pointers: `Cannot create a function pointer to "Foo::bar"; it is abstract`.

Don't need to worry about `self::foo` and `parent::foo`, since we throw a class context ambiguity error in those cases. We don't error in the `static::foo` case either.

Reviewed By: periodic1236

Differential Revision: D24004027

fbshipit-source-id: 99aa416b121dc34ee95ee635cd53966517e62791