Allow layout implementation methods to take their own array type
Summary:
We allow layout implementation methods to take their own array type. A
dispatcher presents a weakly-typed interface, invokes the ::As method of the
array we're wrapping, and uses the typed result to invoke the array's
strongly-typed interface. In non-debug builds, we simply perform a
reinterpret_cast to avoid any overhead from this approach.

Differential Revision: D24119377

fbshipit-source-id: 53ff69714b74dece5887dd9572ac793f15087b17