Fix dict/darray appends in ext_curl
Summary:
curl causes dict/darray appends because it stores "easy handle" resources in a darray, appends new elements to it, and removes elements from it.

The order of elements in this darray doesn't seem to matter. Access is tightly controlled; the only things we do are:
1. Remove a Resource from it
2. Iterate over it to find a Resource in find()
3. Set all the handles to executing / to not executing
4. Check if any handle is executing
5. Raise an exception (using the first handle to throw)
6. AwaitAll on the handles

Order doesn't matter for 1-4. Order matters for 5, but since curl is async, which one throws is already non-deterministic. I'm fine changing that. As for 6, that's done with CurlMultiAwait, which iterates over the handles to block on all of them, but doesn't return a map of results or anything. (The main interface for reading this stuff - curl_multi_info_read - returns the result from an arbitrary easy handle, along with the Resource associated with it, so that's iteration-safe.)

As a result, we can implement remove by swapping the value with the last value and popping the vec. If we're worried about the iteration order above, we can also implement it by shifting all the items down, but I don't think that's necessary.

Reviewed By: ricklavoie

Differential Revision: D24130421

fbshipit-source-id: 870df3cfc37044f50bbd2d3f42094d3d5d031117