Fix a few appends in ext_array
Summary:
We're not even close to fixing all the appends in ext_array. array_pad is tricky to convert, because its result depends on whether the input array has elements removed from it or not. array_merge_recursive is a whole other kind of hell. But here are three functions that used to append to dict/darray that no longer do.

The trickiest thing here is the increment-with-overflow on m_nextKI, which I implemented as `nextKI = int64_t(size_t(converted.val().num) + 1)` - is there a better way?

Reviewed By: ricklavoie

Differential Revision: D24130945

fbshipit-source-id: 8b5dd98ac6911d6d79604fae48996a887ed3dae8