Add virtual BespokeArray::SetLegacyArray method
Summary:
We'd like to get rid of the ArrayData::copy API, which must return a refcounted copy. If we can do so, we can ensure that all instances of certain bespoke array subtypes are static (e.g. EmptyVec and EmptyDict).

The copy API is only used by Hack Array Migration. It's used for tagging and marking arrays. We can fold its use into the tagging and marking helpers, instead, giving them the same copy/escalation semantics as any other logical array mutation. In this diff, we make this change for the marking API.

Differential Revision: D24114209

fbshipit-source-id: b58780811389c1d044014d5f04d0853a2e9e96db