Drop copies in recursive mutation helper
Summary:
The trickiest copy to eliminate is this one that's used for this optimized, templated helper that applies an arbitrary mutation to a nested tree of array-likes. We need this helper to be fast in all modes, because it's used for both tagging (which is arrprov-only) and marking (which happens everywhere).

We should be able to make this code generic and speed it up at the same time. To do so, we split out a templated version of this helper that does the fast thing by direct pointer access on PackedArray and MixedArray, then do the generic code via IterateKV. Rather than doing a copy, the generic code just goes straight to the set that we need the copy.

Differential Revision: D24120486

fbshipit-source-id: f67081e55239d6495edf96a19a9619f5253ceb6d