Eliminate ArrayData::copy
Summary: Eliminate the remaining copies. There's one completely unnecessary copy in an extension, which I've cleaned up. All the remaining ones are done for the purpose of setting provenance tags, so we can assume that the tag target is always a (vanilla) dvarray and use either PackedArray::Copy or MixedArray::Copy.

Differential Revision: D24124643

fbshipit-source-id: 2d7ec0724311f0848eee56a64cbe4c18a8d626f5