Make fb303 port binding behavior match that of http-server
Summary:
Binding a port to an AsyncServerSocket has subtly different behavior
depending if you call bind() with a port number directly, versus
resolving the port with a SocketAddress and then calling bind on
that. Namely, it binds to a different ILA address. Match the behavior
of http-server and use a SocketAddress instead of the raw port. While
I'm at it, add an explicit configuration for the address to bind to,
in case we want to override the default behavior.

Reviewed By: mofarrell

Differential Revision: D24157236

fbshipit-source-id: 166138e958c98bec8df24a18af45b4b21c9cfcfa