Do not corrupt stack with SyncReturnBC
Summary:
In the past, any thrown exception would trigger immediate materialization of
all inlined frames on the stack. This opcode was used to aid that process and
to fix frame linkage and return offsets.

This is no longer needed, as unwinder both calls the catch block and tears down
the frame one frame at a time.

What is worse, SyncReturnBC overwrites potentially valid stack values that are
expected to be unwound. As of now, this situation could happen only on a call
from inlined code that fails on an early stack overflow check, which is a rare
situation, but this will be more common as various prologue checks are
reshuffled.

Removal of this opcode also results in a better code, as we no longer stash
addresss of frame pointers that were optimized away on the native stack.

Reviewed By: ricklavoie

Differential Revision: D24157742

fbshipit-source-id: d0b966a180fec01e505ef367e8d46fe221887cf5