Monotype vector implementation
Summary:
An implementation of the MonotypeVec bespoke and its empty seed. EmptyMonotypeVec is used for empty arrays that will escalate into MonotypeVecs as they are populated. Major implementation details:
1) All EmptyMonotypeVecs are static, and escalate to MonotypeVec upon insertion of an element.
2) MonotypeVecs escalate to vanilla if an incompatible element is inserted. They *do not* demote to EmptyMonotypeVec if they are emptied.
3) Currently the type of the MonotypeVec is determined by accessing the layout. This will be optimized in a follow-up (see below).
4) A MonotypeVec consists of the standard ArrayData header, plus a region of storage. The storage consists of consecutive 8-byte values.

This implementation is nearly complete, except for deriving the monotype type from the layout index itself. This is blocked on reserving layout index regions. This implementation also uncovered a flaw in the current Elem{Int/Str} implementation that makes Elem operations unsafe for elements of type ClsMeth. We escalate to vanilla in this case to work around the issue for now.

Differential Revision: D24000781

fbshipit-source-id: 39175a2d7902131c5d64ef918a05a8fa02d1b048