move type parameter State to associate type in SmartConstructors
Summary:
`SmartConstructors` takes a type parameter for state, when `SmartConstructors` is used as constraint in a type, this type also needs to carry a type parameter for the state in `SmartConstructors`. Moving the State type parameter can avoid it.

This saves one type parameter in `Parser`, and in the next diff, it also saves one type parameter in `Lexer`

Differential Revision: D24101168

fbshipit-source-id: d16e1b11bae97bee8fee87db84eb8de2cdffd0ef