Ensure enum types and struct fields are Copy
Summary:
Currently, hh_oxidize has a preference for inlining enum arguments and struct fields rather than referencing them. This helps with cache locality, but hurts memory usage for enum variants (which must all be the size of the largest variant of the same type) and results in moving more bits around in memory (when large structs are passed around by value). If we prefer putting values behind references instead, we can avoid this memory wastage and memmoving, and gain a convenient benefit--we can generally rely upon the fact that all struct fields are Copy.

This is useful for reallocating a struct with changes. For instance, in the direct decl parser, we sometimes need to rewrite the type of a function parameter. Before this change, we needed to do so this way:

```
let convert_param = |param: &'a FunParam<'a>| {
    self.alloc(FunParam {
        type_: self.convert_type(param.type_),
        rx_annotation: param.rx_annotation.clone(),
        ..*param
    })
};
```

Because the `rx_annotation` field is inlined, and not `Copy`, we need to explicitly clone it when using the struct update (`..`) syntax. If we can rely upon all fields being `Copy`, however, we don't have to worry about cloning:

```
let convert_param = |param: &'a FunParam<'a>| {
    self.alloc(FunParam {
        type_: self.convert_type(param.type_),
        ..*param
    })
};
```

I think this will be especially useful if we begin modifying by-ref decls in more use cases. One such use case that will arise in the near future is normalizing positions in decls for the purpose of position-insensitive hashing, for the naming table.

Differential Revision: D24165170

fbshipit-source-id: 7b4d3d52addb0249f8c946ee1a632b2daf2927bc