Fix race in ext_strobelight async signal processing
Summary:
It's possible for the signal handler to fire during `rds::threadInit` while we make a syscall to bind to a numa node through `numa_bind_to` (or during a context switch) causing us to run `onStrobelightSignal()` in a state where `rds::header()` is initialized but `rds::local::init()` has not occurred making `RDSLocal` thread locals invalid.

Add an atomic flag that can be checked to see if all initialization for RDS is complete and use it in ext_strobelight.

Reviewed By: binliu19, oulgen, jano

Differential Revision: D24167935

fbshipit-source-id: 6b7b4726593035a2c33b94e525ec37d68fc7205d