Type body of closure properly, add fast path
Summary:
When a closure has *explicit* coeffect annotations, e.g.

```
function rx_context()[rx]: void {
  $less_permissive = ()[pure] ==> rx_context();
}
```

we need to type the corresponding capability types and insert them into env,
so that the body no longer sees the improperly type capability due to
the capturing behavior of capabilities (which is a consequence of
the neat design decision to model them as local variables).

However, if a closure *lacks* coeffect annotations, it is redundant
to do this work beacuse there is no semantic difference between redeclaring
the local capability variable with the same type as in the enclosing scope.
This would have a significant performance for deeply nested closures. E.g.,

```
function rx_context()[rx]: void {
   $rx_closure_fast = () ==> {
     // let $#capability of type `rx` be captured if needed
     // (do not insert a new capability into the environment)
   };
   // is the same as, but more performant than
   $rx_closure_slow = ()[rx] ==> {
     // shadow $#capability with a new binding with the same type
   };
}

Reviewed By: DavidSnider

Differential Revision: D24180807

fbshipit-source-id: d10aadea7d53663794c6d40b157ad86631b27293