make the ifc analysis work on batches
Summary:
In the context of the coverage testing I would like to run the IFC analysis
on many files without paying a big cost in startup latency. The solution is
to enable the IFC mode of hh_single_type_check to process a batch of files.

A header is added before the results to simplify automatic processing.

Reviewed By: madgen, zilberstein

Differential Revision: D24169572

fbshipit-source-id: 932f18dd00191094fc1907e245935b47182463d0