Avoid provenance messages in hphp/test/run.php
Summary: When provenance is turned on for cli mode for a devserver (D24167464), the test runner itself is subject to this config and was spewing notices due to the serialization of varrays in `Queue::sendMessage`. These serialized message array is just list-destructured so we can avoid the notices by switching to vecs.

Differential Revision: D24194460

fbshipit-source-id: 2f5ebd22fccc61767b8feffe25b9f5ffb322b843