Simplify function body caching
Summary:
There's no point trying to enter VM at the funcBodyHelperThunk stub, as it just
tries to do a complex work to resync registers to call funcBodyHelper() C++
helper that determines the real address to call. The only benefit this logic
provides is saving a single null check in case the address was already cached.

Since it is used only by enterVMAtFunc(), which is relatively cold and already
performs a lot of more complex work, just call the underlying getFuncBody()
directly. This function alrady tries to use the cached value first, so we will
be losing at most one nullptr check on the hot path of the cold enterVMAtFunc().

Removes one of the last two uses of ActRec::m_numArgs.

Reviewed By: oulgen

Differential Revision: D23416675

fbshipit-source-id: 73b214fee6bdb7e55341addcb4289cf44fabc695