indent for shape_class_constant.php
Summary:
This test was subtly wrong in a way that was cancelled out by a bug. I'm fixing up the test now, and will fix up the bug in the next diff.

Shapes are required to either have all string-literals or all class-const keys. The code picks a *witness*, the lexically first key of the shape as "witness", checks whether it's string-literal or class-const, and requires all other keys to be the same. The code's in Typing.check_shape_keys_validity
```
  let cmp_keys x y = Pos.compare (shape_field_pos x) (shape_field_pos y) in
  let keys = List.sort ~compare:cmp_keys keys in
  match keys with
  | [] -> env
  | witness :: rest_keys ->
    let (env, pos, info) = get_field_info env witness in
    List.fold_left ~f:(check_field pos info) ~init:env rest_keys
```

Now in this test case, lines 4-6 were correctly picking line 4 as the witness and so requiring the first shape to have string-literals keys. That was fine.

But in lines 13-15 it was somehow [bug] picking the lexically LAST key as witness and so requiring this shape to have class-const keys.

The next diff will fix the bug. But for now, I want to change the test so the second shape unambiguously picks the classs-const as witness.

Differential Revision: D24053616

fbshipit-source-id: ba8e2e8c228b180e2961c228cc7dea5cf6839b7a