swallow errors in identifySymbolService
Summary:
The job of identifySymbolService is to find all symbol uses+definitions. To do that, it walks the TAST. Sometimes it has to call into typing, e.g. `Tast_env.referenced_typeconsts` when it look inside an HApply. What should it do if, while working through that, the typing code reports an error? For instance, if it encounters a parameter of type `I<UndefinedClass::Const>` where referenced_typeconsts reports an error on the class being undefined?
1. It could decline to give any answers
2. It could ignore the errors and give the best effort references it can
3. It could crash on the grounds that the reported error isn't being accumulated.

The existing behavior depends on whether the caller did Errors.do_. There is a bewildering mix of conventions, and no clear contract. For instance, TAST computation is always protected by its own Errors.do_. As for IdentifySymbolService, well, if the caller did --find-class-refs and there are fewer than 10 files to check then we are governed by various ad-hoc calls in ServerRpc to Tast_provider.respect_but_quarantine_unsaved_changes which happens (by accident) to do Errors.do_ and discard them. If there are more than 10 files to check then it uses MultiWorker which discards the parent's Errors.do_ and so will crash if it encounters an error.

I think there's a clear simple principled answer to this mess: the contract for identifySymbolService is to make a best-effort attempt to find all the symbols it can, and the presence of typing errors shouldn't cause a crash.

Differential Revision: D24069879

fbshipit-source-id: ee91c9c189137753ddfd4173513210c77277602a