fix pos for self::CONST shape fields
Summary:
Imagine a class C and within it an expression `shape(C::FOO => 3, self::FOO => 4)`.

If we did `hh --find-class-refs C` (which finds uses but not declarations of C) then we'd expect this to find two uses of C, one located at "C::" and one located at "self::".

But the current behavior of hack things that (correctly) there's one use of C located at "C::", and (incorrectly) there's another use of C located at the declaration-site of C.

This also has impact on `hh --refactor Class C D`. Because it thinks there's both a use-site and a declaration-site of C at the line "final class C", then it applies its rename twice to get "final class DD".

This diff fixes the problem by correcting the location of use-site "self::FOO".

Differential Revision: D24042983

fbshipit-source-id: aed6a5855a34ef6b66baf61d9b8288f6bc39a6a6