make --find-class-ref resilient to file move
Summary:
The way find-class-ref and similar tools work is they look at the depgraph "get_deps_set", they fetch the TASTs for each of the specified files, and they see if the class in question is used within any of those TASTs.

What happens if a file has been deleted/moved?

Prior to this diff, we simply raised an exception. Hh_server kept running, but hh_client got the exception back and exited with code 6. In response the `hh` shellscript repeatedly tried again in an infinite loop.

With this diff, if obtaining-TAST gave rise to an exception, and we were unable to read the file from disk/context, then we'll silently skip over the file. After all if the file can't be read off disk then it necessarily can't reference the symbol in question!

Differential Revision: D24075077

fbshipit-source-id: eb5e4f325758ac36148657bf2e50941d129390c9