Disallow new self/parent for classes with reified generics
Summary:
Reified classes do not support `new self` or `new parent` because it isn't possible to specify the generics as argument. There was a task out for a year to fix this, but it was never done. As a concrete example of unsoundness this can lead to see the following example:

```
class C<reify T> {
  public function foo(): C<int> {
    return new self();
  }
}

function breakit(): void {
  $c = new C<string>();
  $c->foo(); // Runtime error
}
```

This was fixed simply by adjusting the existing check we do that bans `new static` for reified classes.

Reviewed By: jamesjwu

Differential Revision: D24231789

fbshipit-source-id: 05e704360f9db2bc5c16c90ba25525e08b791c6a