Drop some reinterpret_cast in MonotypeVec
Summary: Just some style fixes as I was looking at MonotypeVec for "inspiration" (shameless copying) for MonotypeDict. Probably, the fewer places we need to do a reinterpret_cast, the better. const_cast is more restricted, and in the other two places we do it, we don't need casts at all.

Differential Revision: D24235973

fbshipit-source-id: 2e5f20a11be16703278deb8c4efa6f9e80f83d26