Fix HADVAs marked Collection bug
Summary:
new Map($x) implicitly does a dict-cast on $x. new Vector($x) doesn't have any such reasonable behavior, and actually features a subtle HADVAs bug: it clears the legacy mark pre-HADVAs (by creating a new array) but has no effect post-HADVAs. That's the same behavior as casting, but because it isn't actually implemented as a cast, it doesn't log.

We reimplement the Vector constructor with a cast to restore this logging.

Note that both Map and Vector first check whether the input is empty and return early in that case. That means that neither of them log in this case. I'm fine with continuing with this implementation - it's ever so slightly different from a pure cast, but not in a way that will cause issues.

mcolavita, note that we also had a subtle vanilla-bespoke bug here, too! We'd clear the legacy mark for bespoke vecs but not for vanilla vecs. That's also fixed.

Reviewed By: ricklavoie

Differential Revision: D24238444

fbshipit-source-id: 379eb5c2a6ac2b04f793c984e4c4655f437a63bc