Fix CLI server when files are not accessible
Summary:
Removing stream support from the unit loader accidently broke the use
case of the CLI server executing files which aren't accessible by the
server. For this we use a special CLI client stream wrapper. Since the
wrapper support is gone, we'd try to access the paths like an ordinary
file, fail, and then error.

Restore back partial stream support to the unit loader. We only
support "normal file streams", which are just ordinary files, or
proxies to the CLI client.

Differential Revision: D24239072

fbshipit-source-id: b9827b6619baebaec706f866c3e8b28fabc3d37c