Support %{date} placeholder in ShardSalt for config overrides
Summary:
We have a use-case, where we would like to periodically rotate salt for partial rollout in order to get coverage of on scripts, which only run on specific hosts.

Supporting %{date} takes care of it for use. This ensures, that
1. subsequent re-runs of the script within short time window use the same salt, which is helpful for debugging issues
2. salt is rotated daily

I somewhat arbitrary chose YYYY-mm-dd format for the date.

Choice of %{date} placeholder style is consistent with other placeholders, already supported in other parts of config.hdf.

Reviewed By: ricklavoie

Differential Revision: D24238592

fbshipit-source-id: 0569c0aee223fd3556d7cc14dfd68f53d6b3a6cc