Fix spurious failures in tier overrides tests
Summary:
These tests set a RSS limit. However we can easily go over
this limit during init (for example if we need to unpack embedded
data). If we do so, the runtime fails in inscrutable ways (with
std::bad_alloc and other random crashes). This is especially hard to
diagnose because if the hackc binary already exists, or if you run the
tests as a group (and some other test unpacks the embedded data),
everything works. It only fails if you don't have the unpacked data
and you run the test by itself.

Since this is just testing the override mechanism, we don't actually
care about the RSS, so just remove that config. Remove previous
attempt to bandaid this issue by ignoring "failed to init member
reflection" errors (which is a serious issue and shouldn't be ignored
with a regexp).

Reviewed By: jano

Differential Revision: D24245818

fbshipit-source-id: 644f105cfb1c32defdfaa618f5dd567a2f8fe609