implement pipes
Summary:
Lucky for us, Pipe is not just another binary operator.

The implementation is quite simple and I mostly copied it from `typing.ml`. The only slightly tricky bit is that the `$$` is only in scope while typechecking the RHS, so after checking the RHS, we have to restore the previous value of `$$` in the env. Everything else falls out from our existing machinery

Reviewed By: mpu

Differential Revision: D24172445

fbshipit-source-id: 2a3dbff1fc3dc6c7713e5f34f8fcbb695cc5e683