TCRangeInfo: use SrcKey instead of Offset
Summary:
Offset is not able to properly encode non-bytecode positions, such as prologue
and soon function entry. It also relies on the sha1 component to identify the
unit and it's easy to make a mistake and assume all Offsets in a translation
belong to the same Func.

Instead, use SrcKey to properly encode the position.

Also, fixes a number of bugs:
- Eval.ProfileBC now properly records instructions for inlined code
- prologue line numbers now point to the function declaration rather than its
  first bytecode
- some tc-print logic can now distinguish prologues from the first bytecode,
  although the logic is not yet used for prologues

Reviewed By: ricklavoie

Differential Revision: D24242720

fbshipit-source-id: 8ea30c70bf08c5efa8ad9c6145a48c3ec8639c04